fixing camera timestamp

+++ include/cv_connection.hpp
@@ -61,7 +61,7 @@ public:
 
   void WriteToJpeg(uint8_t *data, uint32_t compressed_size, const ros::Time &time_stamp);
 
-  void PubFrameInfo(const ros::Time &time_stamp, uint32_t camera_timestamp);
+  void PubFrameInfo(const ros::Time &time_stamp, uint64_t camera_timestamp);
 
   void check_for_subscribers();
